Rename module and remove comments
