+++ src/sentry/replays/consumers/recording_buffered.py
@@ -61,7 +61,7 @@ from sentry.replays.lib.storage import (
     make_video_filename,
     storage_kv,
 )
-from sentry.replays.usecases.blocklist import mobile_processing_is_blocked
+from sentry.replays.usecases.denylist import mobile_processing_is_blocked
 from sentry.replays.usecases.ingest import decompress, process_headers, track_initial_segment_event
 from sentry.replays.usecases.ingest.dom_index import (
     ReplayActionsEvent,
@@ -226,8 +226,7 @@ def process_message(buffer: RecordingBuffer, message: bytes) -> None:
             logger.exception("Could not decode recording message.")
             return None
 
-    # Sales-led organizations are temporarily blocked during the open beta period
-    # unless they explicitly opt in.
+    # TODO: Temporary open beta block. Remove after GA.
     if mobile_processing_is_blocked(decoded_message):
         return None
